chore: move each chain into separate file
